Reload nginx really gracefully
Introduce `nginx_reload_script` configuration option, allowing to
specify a script which will be used to reload `nginx`.

Change the existing actions in such a way that while reloading `nginx`,
after we sent `SIGUSR2` to `nginx` asking it to launch a new instance of
itself, let's check that the new instance actually launched before
killing the old one.