Only disable the nginx listener when a namespace is Envoy-only
The current behavior is to disable both the haproxy and nginx listener
when a namespace is Envoy-only. There's a race condition when disabling
a namespace where haproxy (the service that listens on a UNIX socket)
reloads before nginx (the service that listens on the proxy port and
proxies to haproxy) which results in 503s being served for about 10
seconds.

The new approach is to just keep the haproxy listener around and only
tear down the nginx listener. This avoids complex coordination that
would be required to ensure they get torn down / brought up in the
correct order to avoid serving errors, and should be effectively
equivalent since nothing will be using the haproxy UNIX socket.

Ticket: MESH-931